agents/langraph-shop-planner/requirements.txt
langchain>=0.1.0
langgraph>=0.0.10
pydantic>=2.5.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.1.0
streamlit>=1.29.0
openai>=1.6.1
chromadb>=0.4.18
numpy>=1.26.0
pandas>=2.1.4
python-multipart>=0.0.6
watchfiles>=0.21.0 